Add dockercfg_base64 variable for tectonic-ui service
More generic intent though: preload all kubernetes nodes with a way to
access private docker images.  For now this is just substituted in to
the tectonic-ui service directly.  Eventual goal is to have this preload
node:~root/.dockercfg with the contents of settings.cluster.dockercfg

eg:
cluster:
  dockercfg:
    # contents of ~/.dockercfg converted to yaml
  # or..
  dockercfg_base64:
    # output of: cat ~/.dockercfg | base64

This caused master's userdata to go over 16KB, so had to refactor what
was in there.  Kraken services are now rendered and created all at once
rather than a unit-per