Remove the wait4node unit, split up common node tasks
Rather than waiting for api server to come up, and the node to get
registered, just keep retrying until it works.  Once the label works,
just sleep for 15 minutes.  This is so I can have systemd run this in
the background rather than a blocking one-shot. I'm assuming kubectl
exits successfully if the label is successful, need to test this.