Bypass journald for kubernetes components
Workaround for coreos/bugs#990

Logging to stderr is unbuffered, so whatever is causing systemd-journald
to block / not process messages fast enough eventually causes the kube
components logging to stderr to block

Use a log_dir parameter to have the kubernetes components log directly
to files.  Seemed more flexible than writing a wrapper shell script to
`>> /var/log/k8s/foo 2>&1` although not sure I'm a fan of the way files
are broken out per-level

For now /var/log/k8s is still on the root volume. If we find we're
getting i/o limited there, we should follow up with something that puts
that on a more performant volume.