Ensure services are started during ansible-playbook runs
This looks like (and may be) a lot of redundant copy-pasta. The handlers
are used to restart services when their unit files get modified.  The
tasks are used to ensure services are started even when the unit files
aren't modified.

We run into the latter case when ansible lays down unit files, restarts
systemd, and then for Reasons, dies before the handlers are able to
restart services.  The next run through, the unit files are already
present, so notify/handlers won't fire.  Hence, we define the service
with a state of "started" so it will only be started if it's not already
running.