Prevent k2 from stomping on user files (#170)
* Stop openssh from creating ~/.rnd

openssh creates an entropy file at ${HOME}/.rnd which, in the docker
container, is owned by root. When this file is in a user-owned
directory (or home) it can stomp on the user-owned file by the same
name.

By setting the RANDFILE environment variable, we can force it to use a
safe filename.

* Make openssh create safe .srl files

openssh x509 was being called with CAcreateserial which was causing it
to create `~/.srl`. This specifies a proper file location to prevent
potentially overwriting a user's own .srl file in their home directory.

* Use ${HOME} instead of ~ for consistency and readability

* Stop ansible and terraform from creating files in $HOME

Both ansible and terraform try to create files/directories under `$HOME`
with no way of reconfiguring that behavior. Change the `$HOME` export to
allow them a safe place to store their workfiles.