Don't use temporary directory for $HOME (#182)
* Revert "Stop ansible and terraform from creating files in $HOME"

This reverts commit 8f7ac494c1be8364b00fd6cf43d74fae73a66986.

* Document how to map the needed files instead of $HOME

Add documentation to the README.md for binding just the files we need
for K2 instead of the entire home directory. This prevents the K2
processes from writing potentially overwriting files in the user's home
directory when running as a container.

Fixes #181 and is associated with #97

Similar changes to k2cli should be done,
https://github.com/samsung-cnct/k2cli/issues/51

* Fix extra space in string

* Add default paths for files we want to bind mount

`k2cli` walks this configuration file looking for files to bind mount
into the container. Add sane defaults so they will be found and mounted.

Fixes https://github.com/samsung-cnct/k2cli/issues/51

* Allow boto to use the expected configuration precedence

* Match k2cli function a little more closely.

* retry stuff where the api server isn't listening sometimes