Delete cluster when not fully up (#226)
* Set run-services as a dynamic include

When bringing down a cluster, run-services is still processed and will
fail because the lookup() of admin.kubernetes may fail if the file
doesn't exist. Since we are not actually including this task if we're
bringing down the cluster, we don't need that lookup to succeed. By
marking the include as not static, it won't be processed.

* Add a name so I stop seeing "fail" in the tasks

* Allow "down" even if terraform apply fails

If terraform apply partially fails, a tfstate file will exist but other
files will not. Work around this and allow `down` to run successfully.

* Don't apply filter to entire logic statement