allow a user to specify a pre-existing VPC for aws clusters (#579)
* allow a user to specify a pre-existing VPC for aws clusters
to use a pre-existing VPC a user will need to specify the vpc id,
a route table in that vpc and the id of the default security
group of the vpc.  the user should also be certain they have set
up all of the network CIDRs correctly as K2 will not help with
that.