set aws access/secret if unset and credential file is set (#633)
this is being done at all because K2 allows a user to specify only
the aws credential file, this is sane.  however, the ansible ec2_*
modules do not allow for specifying the credential file in the
module, they rely on setting the access key and secret key directly
or using the environment variables for those.

if all the modules were all using boto3 we could use a single
environment variable that would point to the credential file but
alas, the older ec2_ modules use boto2 which has no environment
variable mechanism for setting the credential file location.

and this is happening prior to each ec2_ module invocation because
I couldn't figure out how to set it globally.  I was unable to find
a filter that could traverse both list and dict variables, combine
seems to only work on dicts.

This is required for k2cli CI and to support users using non-
standrd credential file locations